New GrandExchange class in place
This is the start of a better ge command that responds with results and
data about the item, instead of the boring old link to the website. The
API has a good amount of info, its just that it takes the ID of the item
as the search query (which I will need to figure out). The API can also
respond with data to create a graph.