GE class now searches a JSON list of item ID numbers
This makes it so you can search with item names. As of right now only
full names work, however the next step is to make it work when it
the query contains most of the name (i.e. `rune plate` not
`rune platelegs`).

I'd also like it to possibly say other results that could've been
possible with a query (i.e. `rune plate` could be `rune platelegs` or
`rune platebody`...)

By searching this JSON file it will also mean that any new items added
to the game will have to be added to this file, so its not ideal for
futureproof.