XP tracking command added
!b xp [username] shows weekly activity for the user, this includes total
XP gain, total levels gained, rank change. Also shows the status for top
5 stats trained, and recent boss kills (if any).

I removed the sig command, because this replaces it.

Added urls to urls.py for CML commands.

The CML API is very limiting in that you can't call it too often without
paying for a key, so all the data is scraped off of the website. This
isn't completely ideal because of futureproofness but its working.