Redo of search algorithm for price command
The command first searches the ge API to see if it can find a match,
if it finds an exact match or a substring match then it shows the the
single result. If there are multiple results they are listed with
prices, this makes the command useful to lookup prices on items that
have multiple versions (i.e. potions).

If the API returns nothing, then as a final attempt it will search
the JSON file. This fills a few more edge cases like searching 'whip'.