Tracker no longer uses Hiscore
Since the official hiscores page frequently times out it is more
reliable to get the stats from the CML site. Since the levels from CML
are virtual (above 99), they need to be handled a little differently.

This seems to speed up the command overall because it doesn't have
to wait for the hiscore page's response.

There are two new functions in Tracker, get_lvl which returns the lvl
for a skill by name (this could be >99), and get_non_virtual_lvl which
returns the lvl for a skill by name (the max being 99).