Tracker parses response when attempting to update a user's XP
A status message is added to the footer of the embed. This shows whether
or not the update went successful or not. If the API responds with the
busy signal (-4) then the user is shown a link to manually update the XP
if they know that there is new gains.

Tracker no longer raises the NoDataPoints exception and will instead
show a message describing that there aren't any gains to be shown.

Even if the XP wasn't updated the latest gains are shown if available.